Fix30 and Fix56: * implemented check of VE data types (check for text containing "," or "." and check in keypress event) * changed Hibernate types of numberpu to Integer (AvailArticleInDepot, IncomingArticle and OutgoingArticle including the DTOs of those Classes)

+++ ERP-light/src/at/erp/light/view/dto/AvailableArticleDTO.java
@@ -5,13 +5,13 @@ public class AvailableArticleDTO {
 	
 	private int articleId;
 	private ArticleDTO articleDTO;
-	private double availNumberOfPUs;
+	private Integer availNumberOfPUs;
 	
 	public AvailableArticleDTO() {
 	}
 	
 	public AvailableArticleDTO(int articleId, ArticleDTO articleDTO,
-			double availNumberOfPUs) {
+			Integer availNumberOfPUs) {
 		super();
 		this.articleId = articleId;
 		this.articleDTO = articleDTO;
@@ -34,11 +34,11 @@ public class AvailableArticleDTO {
 		this.articleDTO = articleDTO;
 	}
 
-	public double getAvailNumberOfPUs() {
+	public Integer getAvailNumberOfPUs() {
 		return availNumberOfPUs;
 	}
 
-	public void setAvailNumberOfPUs(double availNumberOfPUs) {
+	public void setAvailNumberOfPUs(Integer availNumberOfPUs) {
 		this.availNumberOfPUs = availNumberOfPUs;
 	}
